requirements.txt
discord.py==2.3.2
requests==2.31.0
openai==1.3.5
airtable-python-wrapper==0.15.3
pillow==10.1.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2